perf: accept new cart when getting discount
BREAKING CHANGE: The functions used for getting discounts for a cart
now accept the new cart object, rather that the cart id. This is
required because the discounts should be calculated based on the items
of the new cart which is in the process of being persisted rather than
the items of the persisted cart.

Signed-off-by: tedraykov <tedraykov@gmail.com>